% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{run_app}
\alias{run_app}
\title{Personal greeting as a Shiny app}
\usage{
run_app()
}
\value{
Shiny app showcasing the personal greeting feature.
}
\description{
Greet a person and appropriately capitalize their name
as a Shiny app.
}
